feat: connect to LocalNet (easily)
1.  Made `synchronizerId` optional in most places.  It is retrieved from the
    validator app if not specified.
2.  Added a `localnet` configuration.
3.  Changed the example dApp to retrieve the Ledger version, and toggle the ping
    contract based on that.

The `...e.network` pattern in the wallet-gateway frontend was giving me trouble
about undefined vs. omitting a field in between the types generated from the
OpenAPI spec and the zod parser.  I tried a few different patterns abusing the
splat operator and filtering out the extra types, but in the end I went with
something more explicit and simple.

Signed-off-by: Jasper Van der Jeugt <jasper.vanderjeugt@digitalasset.com>