fix(wallet-gateway-remote): store all tx params
In case we are dealing with participant signing, we submit the transaction
using `/submit-and-wait`.  In contrast to the prepare/sign workflow of
external signing, this needs all parameters (including, in particular,
`.disclosedContracts`).

This means we need this info stored.  This PR modifies the semantics of
`payload` to include all parameters rather than just `.commands`.

Signed-off-by: Jasper Van der Jeugt <jasper.vanderjeugt@digitalasset.com>