refactor(core-ledger-client): split up token standard service (#982)
The `TokenStandardService` currently depends on the scan client and/or the scan proxy client.

However, in the token standard definition, there is nothing in the token standard that refers to scan APIs.  More practically, I would also like to be able to run the token standard portfolio without the user (or app provider) having to configure scan URLs.

This refactor puts us on the right path for that, it removes the following non-standard concerns into a separate `AmuletService`:

- Tapping
- Transfer preapprovals (this may become part of the token standard later)
- Featured apps

Note that this only affects the core libraries.  We could do a similar split in the Wallet SDK (but that would be a breaking change).

I picked the shortest name, we could also consider:

- `AmuletTokenService`
- `SpliceTokenService`
- `CantonTokenService`
- `CantonCoinService`

Signed-off-by: Jasper Van der Jeugt <jasper.vanderjeugt@digitalasset.com>