feat: add rough portfolio dApp (#1028)
This PR adds a Portfolio dApp.  It is still VERY rough around the edges and should not be used for any real work.

However, getting a working version into `main` would help with collaboration, and hopefully it also means that we can do smaller subsequent PRs from now on.

Things that work:

- Settings token registries
- Tapping (amulet only)
- Two-step transfers
- Transaction history

Things to call out:

- The app adds all React components into an `oldcomponents/` folder.  This is intentional as it will allow @fayi-da to replace these iteratively without conflicts.
- Anything in `services/` should be React-agnostic
- Some parts of the application already use React hooks but more should be refactored in to avoid effects in the individual tabs

Things that need to be addressed (preferably in follow-up PRs, but I'm happy to address the essentials):

- Feature: allocations
- Add more hooks to make this app more idiomatically react (without affecting portability)
- Error handling
- Test cases

Signed-off-by: Jasper Van der Jeugt <jasper.vanderjeugt@digitalasset.com>