feat(example-portfolio): add proper RegistryService (#1030)
This adds a more full-featured registry service to the portfolio dApp.

In addition to managing the URLs, it will now retrieve token information from all registries it knows about.

As a React layer on top, I added `useRegistryUrls` and `useInstruments` to retrieve all of this.

I also added a SelectInstrument component that uses this, and changed tapping and transfers to use this.

Signed-off-by: Jasper Van der Jeugt <jasper.vanderjeugt@digitalasset.com>