feat(example-portfolio): accepting allocations (#1063)
This PR implements the happy path for allocations.  Since the token standard specifies no way to initialize an allocation flow, an external tool or script needs to be used.

I adapted the example script using the wallet SDK to initialize this allocation, then pause so the user can complete the allocation steps in the UI.  After that the script can resume and exercise the allocations atomically.

It's still missing a few parts:

- Accepted allocations don't show as such
- Exercised allocations don't show up in the transaction history

These will be added as separate PRs since this one is already getting sizeable to review.